Inteiros

+++ frequencia.cpp
@@ -2,7 +2,7 @@
 int main(){
     int n=0, c=0, qtd_numeros=1, i = 0, j = 0, k = -1;
     bool achou = false;
-    double v[n], cont[c], num;
+    int v[n], cont[c], num;
 
     do{
         printf("N = ");
@@ -13,7 +13,7 @@ int main(){
     k = -1;
     for(i = 0; i < n; i++) {
         printf("V[%d] = ", i);
-        scanf("%lf", &num);
+        scanf("%d", &num);
         achou = false;
         j = 0;
 
@@ -21,7 +21,7 @@ int main(){
             if (v[j] == num) {
                 achou = true;
                 cont[j] = cont[j]+1;
-                printf("v[%d] == %.1lf\n", j, num);
+                printf("v[%d] == %d\n", j, num);
             } else {
                 j = j+1;
             }
@@ -32,13 +32,13 @@ int main(){
             k = k+1;
             v[k]= num;
             cont[k]=1;
-            printf("achou == false. v[%d]=%.1lf e cont[%d]=1\n",k, num, k);
+            printf("achou == false. v[%d]=%d e cont[%d]=1\n",k, num, k);
         } else {
-            printf("achou == true. v[%d]=%.1lf e cont[%d]=1\n",k, num, k);
+            printf("achou == true. v[%d]=%d e cont[%d]=1\n",k, num, k);
         }
     }
 
     for (j = 0; j<= k; j++) {
-        printf("%.1lf apareceu %.1lf\n", v[j], cont[j]);
+        printf("%d apareceu %d\n", v[j], cont[j]);
     }
 }
